@@ -1,6 +1,7 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: "androidx.navigation.safeargs"
 apply plugin: 'android-apt'
 
 android {
@@ -15,12 +16,12 @@ android {
             checkReleaseBuilds false
         }
     }
-    compileSdkVersion 27
-    buildToolsVersion '28.0.2'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "tk.samgrogan.pulp"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -51,24 +52,29 @@ repositories {
 dependencies {
     ext.anko_version = '0.10.7'
     def nav_version = "1.0.0-alpha06"
+    def lifecycle_version = "1.1.1"
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:27.0.0'
-    implementation 'com.android.support:design:27.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.github.nihad92:SwipeableCards:1.0.19'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha2'
+    implementation 'com.github.Gh0stWires:SwipeableCards-Right-Only-:1.1.1'
     implementation 'com.google.firebase:firebase-core:16.0.4'
     implementation 'com.github.eschao:android-PageFlip:1.0.2'
-    implementation "android.arch.navigation:navigation-fragment:$nav_version" // use -ktx for Kotlin
-    implementation "android.arch.navigation:navigation-ui:$nav_version" // use -ktx for Kotlin
+    implementation 'org.koin:koin-android-viewmodel:1.0.1'
+    implementation "android.arch.navigation:navigation-fragment-ktx:$nav_version" // use -ktx for Kotlin
+    implementation "android.arch.navigation:navigation-ui-ktx:$nav_version" // use -ktx for Kotlin
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.0.0' // use -ktx for Kotlin
+    implementation 'androidx.lifecycle:lifecycle-livedata:2.0.0'
+    implementation 'androidx.lifecycle:lifecycle-runtime:2.0.0'
     wearApp project(':pulpremote')
     implementation "org.jetbrains.anko:anko-common:$anko_version"
     implementation project(':library')
-    implementation 'com.google.android.gms:play-services-wearable:16.0.1'
-    implementation 'com.google.android.support:wearable:2.0.2'
     implementation 'com.firebaseui:firebase-ui-auth:4.0.0'
-    implementation 'com.android.support:recyclerview-v7:27.0.0'
-    implementation 'com.android.support:support-v4:27.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.google.android.gms:play-services-auth:16.0.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha2'
     implementation 'com.google.firebase:firebase-database:16.0.3'
     implementation 'com.daprlabs.aaron:cardstack:0.3.1-beta0'
     implementation 'com.google.firebase:firebase-auth:16.0.4'
