@@ -1,15 +1,22 @@
 package tk.samgrogan.pulp.adapters
 
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
+import tk.samgrogan.pulp.ui.reader.PageFragment
 
 class ComicPageAdapter(fragmentManager: FragmentManager): FragmentPagerAdapter(fragmentManager) {
+    private lateinit var pages: MutableList<Int>
+
     override fun getItem(position: Int): Fragment {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+        return PageFragment.newInstance(pages[position])
     }
 
     override fun getCount(): Int {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+        return pages.size
+    }
+
+    fun setPages(pages: MutableList<Int>) {
+        this.pages = pages
     }
 }
