@@ -7,9 +7,9 @@ import android.graphics.Bitmap;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.annotation.Nullable;
-import android.support.design.widget.BaseTransientBottomBar;
-import android.support.design.widget.Snackbar;
+import androidx.annotation.Nullable;
+import com.google.android.material.snackbar.BaseTransientBottomBar;
+import com.google.android.material.snackbar.Snackbar;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -28,8 +28,8 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
-import tk.samgrogan.pulp.Data.ReadCBR;
-import tk.samgrogan.pulp.Data.ReadCBZ;
+import tk.samgrogan.pulp.data.ReadCBR;
+import tk.samgrogan.pulp.data.ReadCBZ;
 import tk.samgrogan.pulp.Models.ComicDataObject;
 import tk.samgrogan.pulp.Models.Comics;
 import tk.samgrogan.pulp.R;
