@@ -2,11 +2,10 @@ package tk.samgrogan.pulp.ui.reader
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
 import tk.samgrogan.pulp.R
 
 class PageFragment : Fragment() {
@@ -15,6 +14,16 @@ class PageFragment : Fragment() {
                               savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_page, container, false)
+    }
+
+    companion object {
+        fun newInstance(page: Int): PageFragment {
+            val fragment = PageFragment()
+            val args = Bundle()
+            args.putInt("someInt", page)
+            fragment.arguments = args
+            return fragment
+        }
     }
 
 }
