@@ -2,8 +2,8 @@ package tk.samgrogan.pulpremote;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.support.wearable.view.WatchViewStub;
 import android.util.Log;
 import android.view.View;
