@@ -53,27 +53,27 @@ public class RarCRC {
 		// #if defined(LITTLE_ENDIAN) && defined(PRESENT_INT32) &&
 		// defined(ALLOW_NOT_ALIGNED_INT)
 		/*
-		for (int i = 0; (0 < size) && i < Data.length - 8
-				&& ((Data[i + 8] & 7) != 0); i++) {
-			startCrc = crcTab[(short) (startCrc ^ Data[i]) & 0x00FF] ^ (startCrc >>> 8);
+		for (int i = 0; (0 < size) && i < data.length - 8
+				&& ((data[i + 8] & 7) != 0); i++) {
+			startCrc = crcTab[(short) (startCrc ^ data[i]) & 0x00FF] ^ (startCrc >>> 8);
 			size--;
 		}
 		
 		for (int i = 0; size >= 8; i += 8) {
-			startCrc ^= Data[i + 0] << 24;
-			startCrc ^= Data[i + 1] << 16;
-			startCrc ^= Data[i + 2] << 8;
-			startCrc ^= Data[i + 3];
+			startCrc ^= data[i + 0] << 24;
+			startCrc ^= data[i + 1] << 16;
+			startCrc ^= data[i + 2] << 8;
+			startCrc ^= data[i + 3];
 
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
 
-			startCrc ^= Data[i + 4] << 24;
-			startCrc ^= Data[i + 5] << 16;
-			startCrc ^= Data[i + 6] << 8;
-			startCrc ^= Data[i + 7];
+			startCrc ^= data[i + 4] << 24;
+			startCrc ^= data[i + 5] << 16;
+			startCrc ^= data[i + 6] << 8;
+			startCrc ^= data[i + 7];
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
 			startCrc = crcTab[(short) startCrc & 0x00FF] ^ (startCrc >>> 8);
@@ -84,11 +84,11 @@ public class RarCRC {
 		for (int i = 0; i < size; i++)
 		{
 /*
-			// (byte)(StartCRC^Data[I])
+			// (byte)(StartCRC^data[I])
 			int pos = 0; // pos=0x00000000
 			pos |= startCrc; // pos=ffffffff
 			
-			pos ^= Data[i]; // Data[0]=0x73=115dec --> pos=140
+			pos ^= data[i]; // data[0]=0x73=115dec --> pos=140
 			System.out.println(Integer.toHexString(pos));
 			
 			// Only last 8 bit because CRCtab has length 256
@@ -113,7 +113,7 @@ public class RarCRC {
 			//System.out.println(Integer.toHexString(startCrc));
 			
 			// Original code:
-			//StartCRC=CRCTab[(byte)(StartCRC^Data[I])]^(StartCRC>>8);
+			//StartCRC=CRCTab[(byte)(StartCRC^data[I])]^(StartCRC>>8);
 		}
 		return (startCrc);
 	}
@@ -130,14 +130,14 @@ public class RarCRC {
 //	public static void main(String[] args)
 //	{
 //		RarCRC rc = new RarCRC();
-//		//byte[] Data = { 0x72, 0x21, 0x1A, 0x07, 0x00};
+//		//byte[] data = { 0x72, 0x21, 0x1A, 0x07, 0x00};
 //		
-//		byte[] Data = {0x73 ,0x00 ,0x00 ,0x0D ,0x00 ,0x00 ,0x00 ,0x00 ,0x00 ,0x00 ,0x00};
+//		byte[] data = {0x73 ,0x00 ,0x00 ,0x0D ,0x00 ,0x00 ,0x00 ,0x00 ,0x00 ,0x00 ,0x00};
 //		
 //		int crc = 0x90CF;
 //		
 //
-//		int result = rc.checkCrc(0xFFFFffff, Data,0,Data.length);
+//		int result = rc.checkCrc(0xFFFFffff, data,0,data.length);
 //		System.out.println("3: "+Integer.toHexString(~result&0xffff));
 //		
 //	}
